small updates

+++ stocks/stocks.py
@@ -1,4 +1,4 @@
-#!./twitterenv/bin/python3
+#!usr/bin/env python3
 from functools import wraps
 from datetime import datetime
 from time import sleep
@@ -12,6 +12,8 @@ import tweepy as tpy
 auth = tpy.OAuthHandler(credentials['consumer_key'], credentials['consumer_key_secret'])
 auth.set_access_token(credentials['access_token'], credentials['access_token_secret'])
 api = tpy.API(auth)
+
+class CommandError(Exception): pass
 
 def listgen(some_gen):
 	@wraps(some_gen)
@@ -74,29 +76,37 @@ def start_listening(delay=0, attributes={'max_filesize': 25 * (1024 ** 3)}):
 	stream.filter(track=stock_symbols())
 
 pid_file = '.pid_python'
+command = None
+command_successful = True
 
 import atexit
+@atexit.register
 def exit_handler():
-	try:
-		os.remove(pid_file)
-	except: pass
+	if command_successful:
+		try:
+			os.remove(pid_file)
+		except: pass
 	print('Completed at %s.' % str(datetime.now()))
-atexit.register(exit_handler)
 
 if __name__ == '__main__':
 	import argparse
 	parser = argparse.ArgumentParser()
 	parser.add_argument('action', nargs=None, default=None, help='Perform action--"start", "stop", or "test"')
 	args = parser.parse_args()
-	if args.action:
-		if args.action == 'start':
-			if os.path.exists(pid_file): raise Exception('Process already running.')
+	command = args.action
+	if command:
+		if command == 'start':
+			if os.path.exists(pid_file):
+				command_successful = False 
+				raise CommandError('Process already running.')
 			with open(pid_file, 'w+') as f: f.write(str(os.getpid()))
 			start_listening()
-		elif args.action == 'stop':
+		elif command == 'stop':
 			try:
 				with open(pid_file) as f: os.kill(int(f.read()), 1)
 			except FileNotFoundError:
-				raise Exception('No process currently running.')
-		elif args.action == 'test':
+				command_successful = False
+				raise CommandError('No process currently running.')
+		elif command == 'test':
 			start_listening(attributes={'max_responses': 1})
+			print('Test Completed.')
